added permission handling

+++ fullstack/vagrant/catalog/Catalog-Final-Project/methods/category/deleteCategory.py
@@ -1,6 +1,6 @@
 from flask import render_template, url_for, flash, request, redirect
 from methods import routes
-from helpers import sessionMaker
+from helpers import sessionMaker, permissions
 from database_setup import Category
 
 session = sessionMaker.newSession()
@@ -13,14 +13,18 @@ session = sessionMaker.newSession()
 @routes.route('/category/<int:category_id>/delete',
               methods=['GET', 'POST'])
 def deleteCategory(category_id):
-    category = session.query(
-        Category).filter_by(id=category_id).one()
+    if permissions.LoggedIn() == True:
+        category = session.query(
+            Category).filter_by(id=category_id).one()
 
-    if request.method == 'POST':
-        session.delete(category)
-        session.commit()
-        flash("Category deleted")
-        return redirect(url_for('routes.viewAllCategories'))
+        if request.method == 'POST':
+            session.delete(category)
+            session.commit()
+            flash("Category deleted")
+            return redirect(url_for('routes.viewCategoryAll'))
 
+        else:
+            return render_template('/category/deleteCategory.html', category=category)
     else:
-        return render_template('/category/deleteCategory.html', category=category)
+        flash("Please login to create a category")
+        return redirect(url_for('routes.viewCategoryAll'))
