+++ fullstack/vagrant/catalog/Catalog-Final-Project/methods/job/deleteJob.py
@@ -6,6 +6,7 @@ from flask import render_template, url_for, flash, request, redirect
 from methods import routes  # Blueprint routes from __init__.py
 from helpers import sessionMaker
 from database_setup import Job
+from helpers import permissions
 
 session = sessionMaker.newSession()
 
@@ -22,12 +23,22 @@ On POST request: deletes a menu job and updates user.
 def deleteJob(employer_id, job_id):
     i = session.query(Job).filter_by(id=job_id).one()
     if request.method == 'POST':
-        session.delete(i)
-        session.commit()
-        flash("Job? What  job? Job deleted.")
-        return redirect(url_for('routes.viewEmployer', employer_id=employer_id))
+        if permissions.EmployerAdminAndLoggedIn(employer_id) == True:
+            session.delete(i)
+            session.commit()
+            flash("Job? What  job? Job deleted.")
+            return redirect(url_for('routes.viewEmployer',
+                                    employer_id=employer_id))
+        else:
+            flash("Please login.")
+            return redirect(url_for('routes.viewEmployer',
+                                    employer_id=employer_id))
     else:
-        return render_template('/job/deletejob.html', job=i)
+        if permissions.EmployerAdminAndLoggedIn(employer_id) == True:
+            return render_template('/job/deletejob.html', job=i)
+        else:
+            flash("Please login.")
+            return redirect(url_for('routes.viewEmployer', employer_id=employer_id))
 
 
 # Potential future functions
